Add new PantsMainRunConfigurationProducer that overrides the modules of a run configuration to the "main" address.
Our modules are structured to define the JvmBinary in the "main" target. During Run in IntelliJ, it will pick-up the
module path of source code, which usually is not the "main" target. When creating a new Run template, check if we have
a "main" target for the give module address and use it instead. This still allows the developer to change the actual
target of the generated run configuration afterwards.